+++ packages/app-store/routing-forms/env.d.ts
@@ -1,1 +1,1 @@
-declare module "react-awesome-query-builder/lib/config/basic";
+declare module "@react-awesome-query-builder/core";
