+++ packages/app-store/routing-forms/package.json
@@ -7,9 +7,10 @@
   "description": "It would allow a booker to connect with the right person or choose the right event, faster. It would work by taking inputs from the booker and using that data to route to the correct booker/event as configured by Cal user  ",
   "dependencies": {
     "@calcom/lib": "*",
+    "@react-awesome-query-builder/core": "^6.6.13",
+    "@react-awesome-query-builder/ui": "^6.6.13",
     "dotenv": "^16.3.1",
-    "json-logic-js": "^2.0.2",
-    "react-awesome-query-builder": "^5.1.2"
+    "json-logic-js": "^2.0.2"
   },
   "scripts": {
     "lint": "eslint .",
