+++ packages/app-store/routing-forms/types/types.d.ts
@@ -1,5 +1,5 @@
 import type { App_RoutingForms_Form } from "@prisma/client";
-import type { ImmutableTree, Config } from "react-awesome-query-builder";
+import type { ImmutableTree, Config } from "@react-awesome-query-builder/core";
 import type z from "zod";
 
 import type { AttributeType } from "@calcom/prisma/client";
