+++ packages/platform/atoms/package.json
@@ -71,8 +71,9 @@
     "tailwindcss-animate": "^1.0.6"
   },
   "peerDependencies": {
+    "@react-awesome-query-builder/core": "^6.6.13",
+    "@react-awesome-query-builder/ui": "^6.6.13",
     "react": "^18.0.0 || ^19.0.0",
-    "react-awesome-query-builder": "^5.1.2",
     "react-dom": "^18.0.0 || ^19.0.0"
   }
 }
